requirements.txt
pandas>=1.5.0
numpy>=1.21.0
mysql-connector-python>=8.0.0
pyinstaller>=5.0.0 